refactor: move getTopic weights and getDifficulty weights to helper functions in different file
