+++ src/types/interview.ts
@@ -1,4 +1,6 @@
-import { InterviewType } from "@prisma/client";
+import { DifficultyEnum } from "@/constants/difficulties";
+import { Company, InterviewType, Topic } from "@prisma/client";
+import { Nullable } from "./util";
 
 export interface Interview {
   code: string;
@@ -16,4 +18,21 @@ export interface Interview {
   userId: number;
   question: { difficulty: 1 | 2 | 3; title: string };
   type: InterviewType;
+}
+
+// For generating recommendations
+export interface InterviewWithFeedback {
+  updatedAt: Date;
+  questionNumber: number;
+  question: {
+    topics: Topic[];
+    difficulty: DifficultyEnum;
+    companies: Company[];
+  };
+  feedback: Nullable<{
+    id: string;
+    interviewId: string;
+    feedback: string;
+    feedbackNumber: number;
+  }>;
 }
